Clean up spring-* code samples
