+++ hazelcast-integration/pom.xml
@@ -47,14 +47,10 @@
         <module>mongodb</module>
         <module>openshift</module>
         <module>pcf-integration</module>
-        <module>spring-cache-manager</module>
         <module>spring-configuration</module>
         <module>spring-data-hazelcast-chemistry-sample</module>
         <module>spring-data-jpa-hazelcast-migration</module>
         <module>spring-hibernate-2ndlevel-cache</module>
-        <module>spring-jcache</module>
-        <module>spring-security</module>
-        <module>spring-transaction-manager</module>
         <module>springaware-annotation</module>
         <module>springboot-caching-jcache</module>
         <module>springboot-eureka-partition-groups</module>
