--- hazelcast-integration/spring-cache-manager/src/main/java/com/hazelcast/spring/cache/CachingConfiguration.java
@@ -1,36 +1,0 @@
-package com.hazelcast.spring.cache;
-
-import com.hazelcast.client.HazelcastClient;
-import com.hazelcast.client.config.ClientConfig;
-import com.hazelcast.core.HazelcastInstance;
-import org.springframework.cache.CacheManager;
-import org.springframework.cache.annotation.CachingConfigurer;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.cache.interceptor.KeyGenerator;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-
-@Configuration
-@EnableCaching
-class CachingConfiguration implements CachingConfigurer {
-
-    @Bean
-    public CacheManager cacheManager() {
-        ClientConfig config = new ClientConfig();
-        config.getNetworkConfig().addAddress("127.0.0.1:5701");
-        HazelcastInstance client = HazelcastClient.newHazelcastClient(config);
-
-        return new HazelcastCacheManager(client);
-    }
-
-    @Bean
-    public KeyGenerator keyGenerator() {
-        return null;
-    }
-
-    @Bean
-    public PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
-        return new PropertySourcesPlaceholderConfigurer();
-    }
-}
