--- hazelcast-integration/spring-cache-manager/src/main/java/com/hazelcast/spring/cache/DummyBean.java
@@ -1,10 +1,0 @@
-package com.hazelcast.spring.cache;
-
-public class DummyBean implements IDummyBean {
-
-    @Override
-    public String getCity() {
-        System.out.println("getCity called");
-        return "Istanbul";
-    }
-}
