--- hazelcast-integration/spring-cache-manager/src/main/java/com/hazelcast/spring/cache/IDummyBean.java
@@ -1,12 +1,0 @@
-package com.hazelcast.spring.cache;
-
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.stereotype.Component;
-
-@Component("dummyBean")
-interface IDummyBean {
-
-    @Cacheable("city")
-    String getCity();
-}
-
