--- hazelcast-integration/spring-cache-manager/src/main/resources/applicationContext.xml
@@ -1,28 +1,0 @@
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:hz="http://www.hazelcast.com/schema/spring"
-       xmlns:cache="http://www.springframework.org/schema/cache"
-       xsi:schemaLocation=
-               "http://www.springframework.org/schema/beans
-              http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-              http://www.hazelcast.com/schema/spring
-              http://www.hazelcast.com/schema/spring/hazelcast-spring-3.11.xsd
-              http://www.springframework.org/schema/cache
-              http://www.springframework.org/schema/cache/spring-cache.xsd
-       ">
-
-    <cache:annotation-driven cache-manager="cacheManager"/>
-
-    <hz:client id="client">
-        <hz:group name="grp" password="grp-pass"/>
-        <hz:network>
-            <hz:member>127.0.0.1:5701</hz:member>
-        </hz:network>
-    </hz:client>
-
-    <bean id="cacheManager" class="com.hazelcast.spring.cache.HazelcastCacheManager">
-        <constructor-arg ref="client"/>
-    </bean>
-
-    <bean id="dummyBean" class="com.hazelcast.spring.cache.DummyBean"/>
-</beans>
