--- hazelcast-integration/spring-configuration/src/main/java/com/hazelcast/springconfiguration/TestBean.java
@@ -1,15 +1,0 @@
-package com.hazelcast.springconfiguration;
-
-@SuppressWarnings("unused")
-public class TestBean {
-
-    private String result;
-
-    public String getResult() {
-        return result;
-    }
-
-    public void setResult(String result) {
-        this.result = result;
-    }
-}
