--- hazelcast-integration/spring-jcache/src/main/java/com/hazelcast/spring/jcache/IDummyBean.java
@@ -1,9 +1,0 @@
-package com.hazelcast.spring.jcache;
-
-import javax.cache.annotation.CacheResult;
-
-interface IDummyBean {
-
-    @CacheResult(cacheName = "city")
-    String getCity();
-}
