--- hazelcast-integration/spring-jcache/src/main/resources/applicationContext.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright (c) 2008-2018, Hazelcast, Inc. All Rights Reserved.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:hz="http://www.hazelcast.com/schema/spring"
-       xmlns:cache="http://www.springframework.org/schema/cache"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans
-		http://www.springframework.org/schema/beans/spring-beans-2.5.xsd
-		http://www.springframework.org/schema/cache
-		http://www.springframework.org/schema/cache/spring-cache.xsd
-		http://www.hazelcast.com/schema/spring
-		http://www.hazelcast.com/schema/spring/hazelcast-spring-3.11.xsd">
-
-    <cache:annotation-driven/>
-
-    <hz:hazelcast id="instance">
-        <hz:config>
-            <hz:group name="grp" password="pass"/>
-            <hz:network port="5701" port-auto-increment="false">
-                <hz:join>
-                    <hz:multicast enabled="false"/>
-                    <hz:tcp-ip enabled="true">
-                        <hz:member>127.0.0.1:5701</hz:member>
-                    </hz:tcp-ip>
-                </hz:join>
-            </hz:network>
-
-            <!-- cache should be created in instance beforehand -->
-            <hz:cache name="city"/>
-
-        </hz:config>
-    </hz:hazelcast>
-
-    <hz:client id="client" depends-on="instance">
-        <hz:group name="grp" password="pass"/>
-        <hz:network>
-            <hz:member>127.0.0.1:5701</hz:member>
-        </hz:network>
-    </hz:client>
-
-    <hz:cache-manager id="hazelcastJCacheCacheManager" instance-ref="client" name="hazelcastJCacheCacheManager"/>
-
-    <bean id="cacheManager" class="org.springframework.cache.jcache.JCacheCacheManager">
-        <constructor-arg ref="hazelcastJCacheCacheManager"/>
-    </bean>
-
-    <bean id="dummyBean" class="com.hazelcast.spring.jcache.DummyBean"/>
-</beans>
