--- hazelcast-integration/spring-security/src/main/java/com/hazelcast/spring/security/sample/HomeController.java
@@ -1,33 +1,0 @@
-package com.hazelcast.spring.security.sample;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.servlet.ModelAndView;
-
-/**
- * Handles requests for the application home page.
- */
-@Controller
-@SuppressWarnings("unused")
-public class HomeController {
-
-    @RequestMapping(value = {"/", "/welcome**"}, method = RequestMethod.GET)
-    public ModelAndView welcomePage() {
-        ModelAndView model = new ModelAndView();
-        model.addObject("title", "Spring Security Hello World");
-        model.addObject("message", "This is welcome page!");
-        model.setViewName("hello");
-        return model;
-    }
-
-    @RequestMapping(value = "/admin**", method = RequestMethod.GET)
-    public ModelAndView adminPage() {
-        ModelAndView model = new ModelAndView();
-        model.addObject("title", "Spring Security Hello World");
-        model.addObject("message", "This is protected page!");
-        model.setViewName("admin");
-
-        return model;
-    }
-}
