--- hazelcast-integration/spring-security/src/main/resources/log4j.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE log4j:configuration PUBLIC "-//APACHE//DTD LOG4J 1.2//EN" "log4j.dtd">
-<log4j:configuration xmlns:log4j="http://jakarta.apache.org/log4j/">
-
-    <!-- Appenders -->
-    <appender name="console" class="org.apache.log4j.ConsoleAppender">
-        <param name="Target" value="System.out"/>
-        <layout class="org.apache.log4j.PatternLayout">
-            <param name="ConversionPattern" value="%-5p: %c - %m%n"/>
-        </layout>
-    </appender>
-
-    <!-- Application Loggers -->
-    <logger name="com.example.test">
-        <level value="info"/>
-    </logger>
-
-    <!-- 3rdparty Loggers -->
-    <logger name="org.springframework.core">
-        <level value="info"/>
-    </logger>
-
-    <logger name="org.springframework.beans">
-        <level value="info"/>
-    </logger>
-
-    <logger name="org.springframework.context">
-        <level value="info"/>
-    </logger>
-
-    <logger name="org.springframework.web">
-        <level value="info"/>
-    </logger>
-
-    <!-- Root Logger -->
-    <root>
-        <priority value="warn"/>
-        <appender-ref ref="console"/>
-    </root>
-
-</log4j:configuration>
