--- hazelcast-integration/spring-security/src/main/webapp/WEB-INF/hazelcast.xml
@@ -1,6 +1,0 @@
-<hazelcast
-        xsi:schemaLocation="http://www.hazelcast.com/schema/config
-        http://www.hazelcast.com/schema/config/hazelcast-config-3.11.xsd"
-        xmlns="http://www.hazelcast.com/schema/config"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-</hazelcast>
