--- hazelcast-integration/spring-security/src/main/webapp/WEB-INF/spring/spring-security.xml
@@ -1,23 +1,0 @@
-<beans:beans xmlns="http://www.springframework.org/schema/security"
-             xmlns:beans="http://www.springframework.org/schema/beans"
-             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-             xsi:schemaLocation="http://www.springframework.org/schema/beans
-	http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-	http://www.springframework.org/schema/security
-	http://www.springframework.org/schema/security/spring-security-3.2.xsd">
-
-    <http auto-config="true">
-        <intercept-url pattern="/admin**" access="ROLE_USER"/>
-        <session-management>
-            <concurrency-control max-sessions="1"/>
-        </session-management>
-    </http>
-
-    <authentication-manager>
-        <authentication-provider>
-            <user-service>
-                <user name="user" password="password" authorities="ROLE_USER"/>
-            </user-service>
-        </authentication-provider>
-    </authentication-manager>
-</beans:beans>
