--- hazelcast-integration/spring-security/src/main/webapp/WEB-INF/views/hello.jsp
@@ -1,7 +1,0 @@
-<%@page session="false"%>
-<html>
-<body>
-	<h1>Title : ${title}</h1>	
-	<h1>Message : ${message}</h1>	
-</body>
-</html>
